1.12.7.13

+++ modules/core/src/main/java/dev/drawethree/ultraprisoncore/gangs/commands/impl/value/GangValueRemoveSubCommand.java
@@ -33,7 +33,7 @@ public final class GangValueRemoveSubCommand extends GangSubCommand {
 					return false;
 				}
 
-				int amount = Integer.parseInt(args.get(1));
+				long amount = Long.parseLong(args.get(1));
 				String operation = "remove";
 
 				return this.command.getPlugin().getGangsManager().modifyValue(sender, gang.get(), amount, operation);
